moves around model inspector components
