+++ src/lib/components/column-profile/ColumnProfileNavEntry.svelte
@@ -61,7 +61,7 @@
 </div>
 
 <div>
-  {#if sortedProfile && sortedProfile.length && head.length}
+  {#if sortedProfile}
     {#each sortedProfile as column (column.name)}
       <ColumnProfile
         {indentLevel}
