+++ src/routes/_surfaces/inspector/index.svelte
@@ -9,7 +9,7 @@
   import Portal from "$lib/components/Portal.svelte";
   import { drag } from "$lib/drag";
   import { getContext } from "svelte";
-  import ModelInspector from "./ModelInspector.svelte";
+  import ModelInspector from "./model/ModelInspector.svelte";
   import SourceInspector from "./SourceInspector.svelte";
 
   const store = getContext("rill:app:store") as ApplicationStore;
